Admin Digital Panel Module code added
