+++ test-output/Default suite/Default test.xml
@@ -1,15 +1,34 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite hostname="eon-swapnili" name="Default test" tests="1" failures="1" timestamp="26 Mar 2024 12:03:04 GMT" time="38.665" errors="0">
-  <testcase name="TS002_VerifyDigitalAdminPanelSortByOptionsFunctinalityTest" time="20.411" classname="sitecore_PROD.ShowSpecialsApprovals">
-    <failure type="java.lang.AssertionError" message=" List Should be sorted: Lists differ at element [5]: andmore test company != anne mcgilvray &amp;amp; company expected [andmore test company] but found [anne mcgilvray &amp;amp; company]">
-      <![CDATA[java.lang.AssertionError:  List Should be sorted: Lists differ at element [5]: andmore test company != anne mcgilvray & company expected [andmore test company] but found [anne mcgilvray & company]
-	at org.testng.Assert.fail(Assert.java:94)
-	at org.testng.Assert.failNotEquals(Assert.java:513)
-	at org.testng.Assert.assertEqualsImpl(Assert.java:135)
-	at org.testng.Assert.assertEquals(Assert.java:568)
-	at resources.Utility.SortByDigitalIdentity(Utility.java:527)
-	at sitecore_PROD.ShowSpecialsApprovals.TS002_VerifyDigitalAdminPanelSortByOptionsFunctinalityTest(ShowSpecialsApprovals.java:82)
+<testsuite hostname="eon-swapnili" name="Default test" tests="1" failures="1" timestamp="26 Mar 2024 12:46:57 GMT" time="28.669" errors="0">
+  <testcase name="TS002_VerifyDigitalAdminPanelSortByOptionsFunctinalityTest" time="11.35" classname="sitecore_PROD.ShowSpecialsApprovals">
+    <failure type="org.openqa.selenium.ElementClickInterceptedException" message="element click intercepted: Element &amp;lt;button class=&amp;quot;dropdown-toggle-sort btn&amp;quot; type=&amp;quot;button&amp;quot; onclick=&amp;quot;toggleDropdown(&amp;apos;dropdown-sort&amp;apos;)&amp;quot;&amp;gt;...&amp;lt;/button&amp;gt; is not clickable at point (396, 16). Other element would receive the click: &amp;lt;div class=&amp;quot;col-md-6&amp;quot;&amp;gt;...&amp;lt;/div&amp;gt;
+  (Session info: chrome=123.0.6312.58)
+Build info: version: &amp;apos;3.141.59&amp;apos;, revision: &amp;apos;e82be7d358&amp;apos;, time: &amp;apos;2018-11-14T08:17:03&amp;apos;
+System info: host: &amp;apos;EON-SWAPNILI&amp;apos;, ip: &amp;apos;172.22.90.18&amp;apos;, os.name: &amp;apos;Windows 10&amp;apos;, os.arch: &amp;apos;amd64&amp;apos;, os.version: &amp;apos;10.0&amp;apos;, java.version: &amp;apos;1.8.0_311&amp;apos;
+Driver info: org.openqa.selenium.chrome.ChromeDriver
+Capabilities {acceptInsecureCerts: false, browserName: chrome, browserVersion: 123.0.6312.58, chrome: {chromedriverVersion: 123.0.6312.58 (6b4b19e9dfbb..., userDataDir: C:\Users\swapnili\AppData\L...}, fedcm:accounts: true, goog:chromeOptions: {debuggerAddress: localhost:55536}, javascriptEnabled: true, networkConnectionEnabled: false, pageLoadStrategy: normal, platform: WINDOWS, platformName: WINDOWS, proxy: Proxy(), setWindowRect: true, strictFileInteractability: false, timeouts: {implicit: 0, pageLoad: 300000, script: 30000}, unhandledPromptBehavior: dismiss and notify, webauthn:extension:credBlob: true, webauthn:extension:largeBlob: true, webauthn:extension:minPinLength: true, webauthn:extension:prf: true, webauthn:virtualAuthenticators: true}
+Session ID: 5d5109e93bd04acf81eedd6544959893">
+      <![CDATA[org.openqa.selenium.ElementClickInterceptedException: element click intercepted: Element <button class="dropdown-toggle-sort btn" type="button" onclick="toggleDropdown('dropdown-sort')">...</button> is not clickable at point (396, 16). Other element would receive the click: <div class="col-md-6">...</div>
+  (Session info: chrome=123.0.6312.58)
+Build info: version: '3.141.59', revision: 'e82be7d358', time: '2018-11-14T08:17:03'
+System info: host: 'EON-SWAPNILI', ip: '172.22.90.18', os.name: 'Windows 10', os.arch: 'amd64', os.version: '10.0', java.version: '1.8.0_311'
+Driver info: org.openqa.selenium.chrome.ChromeDriver
+Capabilities {acceptInsecureCerts: false, browserName: chrome, browserVersion: 123.0.6312.58, chrome: {chromedriverVersion: 123.0.6312.58 (6b4b19e9dfbb..., userDataDir: C:\Users\swapnili\AppData\L...}, fedcm:accounts: true, goog:chromeOptions: {debuggerAddress: localhost:55536}, javascriptEnabled: true, networkConnectionEnabled: false, pageLoadStrategy: normal, platform: WINDOWS, platformName: WINDOWS, proxy: Proxy(), setWindowRect: true, strictFileInteractability: false, timeouts: {implicit: 0, pageLoad: 300000, script: 30000}, unhandledPromptBehavior: dismiss and notify, webauthn:extension:credBlob: true, webauthn:extension:largeBlob: true, webauthn:extension:minPinLength: true, webauthn:extension:prf: true, webauthn:virtualAuthenticators: true}
+Session ID: 5d5109e93bd04acf81eedd6544959893
+	at sun.reflect.NativeConstructorAccessorImpl.newInstance0(Native Method)
+	at sun.reflect.NativeConstructorAccessorImpl.newInstance(Unknown Source)
+	at sun.reflect.DelegatingConstructorAccessorImpl.newInstance(Unknown Source)
+	at java.lang.reflect.Constructor.newInstance(Unknown Source)
+	at org.openqa.selenium.remote.http.W3CHttpResponseCodec.createException(W3CHttpResponseCodec.java:187)
+	at org.openqa.selenium.remote.http.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:122)
+	at org.openqa.selenium.remote.http.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:49)
+	at org.openqa.selenium.remote.HttpCommandExecutor.execute(HttpCommandExecutor.java:158)
+	at org.openqa.selenium.remote.service.DriverCommandExecutor.execute(DriverCommandExecutor.java:83)
+	at org.openqa.selenium.remote.RemoteWebDriver.execute(RemoteWebDriver.java:552)
+	at org.openqa.selenium.remote.RemoteWebElement.execute(RemoteWebElement.java:285)
+	at org.openqa.selenium.remote.RemoteWebElement.click(RemoteWebElement.java:84)
+	at sitecore_PROD.ShowSpecialsApprovals.TS002_VerifyDigitalAdminPanelSortByOptionsFunctinalityTest(ShowSpecialsApprovals.java:79)
 	at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at sun.reflect.NativeMethodAccessorImpl.invoke(Unknown Source)
 	at sun.reflect.DelegatingMethodAccessorImpl.invoke(Unknown Source)
