+++ test-output/old/Default suite/methods.html
@@ -1,8 +1,8 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="639bd0">  <td>24/03/26 17:32:25</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;ShowSpecialsApprovals.initialize()[pri:0, instance:sitecore_PROD.ShowSpecialsApprovals@39ba5a14]">&gt;&gt;initialize</td> 
+<tr bgcolor="639bd0">  <td>24/03/26 18:16:28</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;ShowSpecialsApprovals.initialize()[pri:0, instance:sitecore_PROD.ShowSpecialsApprovals@39ba5a14]">&gt;&gt;initialize</td> 
 <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1143371233</td>   <td></td> </tr>
-<tr bgcolor="639bd0">  <td>24/03/26 17:32:43</td>   <td>18159</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="ShowSpecialsApprovals.TS002_VerifyDigitalAdminPanelSortByOptionsFunctinalityTest()[pri:2, instance:sitecore_PROD.ShowSpecialsApprovals@39ba5a14]">TS002_VerifyDigitalAdminPanelSortByOptionsFunctinalityTest</td> 
+<tr bgcolor="639bd0">  <td>24/03/26 18:16:45</td>   <td>17229</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="ShowSpecialsApprovals.TS002_VerifyDigitalAdminPanelSortByOptionsFunctinalityTest()[pri:2, instance:sitecore_PROD.ShowSpecialsApprovals@39ba5a14]">TS002_VerifyDigitalAdminPanelSortByOptionsFunctinalityTest</td> 
   <td>main@1143371233</td>   <td></td> </tr>
 </table>
